### 2021-08-05 #3558 @fabian-s
Sorry, I'd rather not -- don't know any Clojure, don't understand what this is supposed to be about.
