### 2021-08-09 #3558 @mjsottile
@Kevin-Mattheus-Moerman I'm not available now, but if this paper still needs an editor when one of the other submissions I'm handling is complete I can do it.  I'm familiar with Clojure and relatives.
