+++ paper.md
@@ -21,23 +21,25 @@ bibliography: paper.bib
 
 # Summary
 
-`diman` (**dim**ensional **an**alysis) is a Clojure based scientific software. It has the following features: create a dimensional formula, create dimensional equation, implement principle of dimensional homogeneity, i.e., perform consistency checks and derive dimensionless products.
+`diman` (**dim**ensional **an**alysis) is a Clojure based scientific software with the ability to: create dimensional formula, create dimensional equation, perform checks for dimensional homogeneity (consistency), and derive dimensionless products.
 
-`diman` provide functions for each step of the analytic process of checking for dimensional homogeneity or deriving dimensionless products. Users can write their custom function that performs the desired process by compounding the functions provided by `diman`. The functions make the computationally repetitive operations hidden. This saves the analyst from laboring in computational tasks while still able to go through the steps of dimensional analysis thereby affording the user to reflect on the analysis.
+`diman` provides functions for each step of the analytic process of checking for dimensional homogeneity or deriving dimensionless products; they make the computationally repetitive operations hidden. Users can write custom compound functions that performs a desired process. This not only saves computational labor, but also can introspect the analysis; the analyst is able to go through the steps of dimensional analysis.
 
 # Statement of need
 
-In general, experiments in science deals with explaining the mechanism of a phenomenon. Description of the mechanism is usually represented by some measurable quantity that addresses the research question such that the value of the quantity is a function of other measurable quantities. Most experimental results suggests many candidates for independent variables of the function. In another scenario, a hypothesis built on experimental results may include a number of independent variables which would be the target for new experiments to test the hypothesis. The independent variables are often dimensional quantities.
+Explaining the mechanism of a phenomenon is often the goal of experiments. As most mechanistic description is expressed in terms of some measurable quantity, its value is a function of other measurable quantities. For example, $F = ma$ where the measurable value of force $F$ is a function of the measurable quantities: mass, $m$; velocity, $v$; and time, $t$.
 
-Dimensionless products are scalars that carries the information of the dimensional quantities that it is a product of. Graphs of dimensionless products often provide more information than graphs having dimensions. Also, points in the dimensionless graph can be determined experimentally. Reducing the number of independent variables to a smaller collection of dimensionless products can assist in understanding the mechanism of the phenomenon [@Langhaar:1951; @Sharma:2021].
+The independent variables of the parent (first or original) function usually has dimensions. Since most of the functions are unknown, the researcher deals with many candidates for independent variable, whose considerations are based on experimental results. Although the mathematical expression of the function is unknown, knowledge of the relationship among the measurable quantities is profitable not only in putting together the series of experimental results to explain the mechanism, but also testing the hypothesis represented by the function.
 
-Numerous software have been developed that deal with dimensions in some shape or form [@Preussner:2018; @Sharma:2021]. Most software incorporate the ability to tag quantities with units. However, few are capable of doing consistency checking and fewer still deal with dimensionless products let alone, deriving dimensionless products.
[…]
