+++ paper.md
@@ -21,17 +21,17 @@ bibliography: paper.bib
 
 # Summary
 
-`diman` (**dim**ensional **an**alysis) is a Clojure based scientific software with the ability to: create dimensional formula, create dimensional equation, perform checks for dimensional homogeneity (consistency), and derive dimensionless products.
+`diman` (**dim**ensional **an**alysis) is a Clojure based scientific software with the ability to: create dimensional formula, create dimensional equation, check dimensional homogeneity (consistency), and derive dimensionless products.
 
-`diman` provides functions for each step of the analytic process of checking for dimensional homogeneity or deriving dimensionless products; they make the computationally repetitive operations hidden. Users can write custom compound functions that performs a desired process. Thus, not only is the computational labor saved, but also introspection of the analysis is possible; the analyst is able to go through the steps of dimensional analysis.
+`diman` provides functions for each step of the analytic process for checking dimensional homogeneity or deriving dimensionless products; they make the computationally repetitive operations hidden. Users can write compound functions that performs a desired process. Thus, not only is the computational labor saved, but also introspection of the analysis is possible; the analyst is able to go through the steps of dimensional analysis.
 
 # Statement of need
 
-Explaining the mechanism of a phenomenon is often the goal of experiments. As most mechanistic description is expressed in terms of some measurable quantity, its value is a function of other measurable quantities. For example, $F = ma$ where the measurable value of force $F$ is a function of the measurable quantities: mass, $m$; velocity, $v$; and time, $t$.
+Explaining the mechanism of a phenomenon is often the goal of experiments. As most mechanistic description is expressible in terms of some measurable quantity, its value is a function of other measurable quantities; the function represents the relationship among the quantities, which provides a mechanistic explanation. For example, $F = ma$ where the measurable value of force $F$ is a function of the measurable quantities: mass, $m$; velocity, $v$; and time, $t$.
 
-The independent variables of the parent (first or original) function usually has dimensions. Since most of the functions are unknown, the researcher deals with many candidates for independent variable, whose considerations are based on experimental results. Although the mathematical expression of the function is unknown, knowledge of the relationship among the measurable quantities is profitable not only in putting together the series of experimental results to explain the mechanism, but also testing the hypothesis represented by the function.
+The independent variables of the parent (first or original) function usually has dimensions. Since most of the functions are unknown, the researcher deals with many candidates for independent variable, whose considerations are based on experimental results. Although the mathematical expression of the function is unknown, knowledge of the relationship among the measurable quantities is profitable not only in putting together the series of experimental results to explain the mechanism, but also testing the hypothesis presented by the function.
 
-If possible, it is beneficial to use the transformed parent function such that all the independent variables are dimensionless. Dimensionless products are scalars that contains information of the dimensional quantities that it is a product of. Not only are points in a graph of dimensionless products experimentally determinable, but also dimensionless graphs can provide more information than dimensional graphs. Reducing the number of independent variables to a smaller collection of dimensionless products can assist in understanding the mechanism of the phenomenon [@Langhaar:1951; @Sharma:2021].
+If possible, it is beneficial to use the transformed parent function, where all the independent variables are dimensionless. Dimensionless products are scalars that contains information of the dimensional quantities that it is a product of. Not only are points in a graph of dimensionless products experimentally determinable, but also dimensionless graphs can provide more information than dimensional graphs [@Langhaar:1951]. Reducing the number of independent variables to a smaller collection of dimensionless products can assist in understanding the mechanism of the phenomenon [@Langhaar:1951; @Sharma:2021].
 
[…]
