### 2021-09-15 #3558 @khinsen
@mjsottile This looks interesting indeed! I'd be happy to review, starting in about a week.
