### 2021-09-17 #3735 @whedon
[REVIEW]: diman: A Clojure package for dimensional analysis
**Submitting author:** @lungsi (<a href="http://orcid.org/0000-0002-1607-0164">B. Lungsi Sharma</a>)
**Repository:** <a href="https://github.com/neuralgraphs/diman" target ="_blank">https://github.com/neuralgraphs/diman</a>
**Version:** v1.1.0
