### 2021-09-30 neuralgraphs/diman#1 @khinsen
Test failures and errors
JOSS reviewer here (from https://github.com/openjournals/joss-reviews/issues/3735). I cloned the repository and ran the tests, but got 2 failures and 6 errors. The full log (output of `lein test` is attached. I used OpenJDK 15 under Ubuntu 20.04 in case it matters.

[test-output.log](https://github.com/neuralgraphs/diman/files/7260926/test-output.log)
