### 2021-10-01 neuralgraphs/diman#1 @lungsi
Thank you @khinsen for pointing out this issue.

I made a minor change in the source code, but I had not updated the test codes. I changed the keys from `:dimension` to `:quantity`, the rationale being that *length* is a quantity and [L] is its dimension.

I have now updated the tests to accommodate the changes and therefore none of the test fails or generate errors.

```
lein test diman.attach-test

lein test diman.buckingham.dimensional-matrix-test

lein test diman.buckingham.dimensionless-product-test

lein test diman.buckingham.homogeneous-equation-test

lein test diman.dimensions-test

lein test diman.filter-test

lein test diman.formula-test

lein test diman.linalg.core-test

lein test diman.linalg.matfun.determinant-test

lein test diman.linalg.matfun.gaussian-elimination-test

lein test diman.linalg.matfun.inverse-test

lein test diman.linalg.matfun.rank-test

lein test diman.linalg.ops.addition-test

lein test diman.linalg.ops.division-test

lein test diman.linalg.ops.multiplication-test

lein test diman.linalg.ops.subtraction-test

Ran 59 tests containing 163 assertions.
0 failures, 0 errors.
```
