+++ README.md
@@ -18,11 +18,11 @@ A Clojure library for applying dimensional analysis.
 
 The easiest way to get all the built-in functions is to be in the default namespace `(in-ns 'diman.default)`. Then, `(println default-functions)` to list all the available functions. 
 
-- Tutorial: Generate dimensional formulae and perform consistency checking; [reStructuredText](./doc/tutorial1.rst), [AsciiDoc](./doc/tutorial1.adoc)
-- Tutorial: Derive dimensionless products; [reStructuredText](./doc/tutorial2.rst), [AsciiDoc](./doc/tutorial2.adoc)
-- Example: Journal Bearing; [reStructuredText](./doc/tutorial3.rst), [AsciiDoc](./doc/tutorial3.adoc)
-- Rationale for the seven base dimensions; [reStructuredText](./doc/rationale1.rst), [AsciiDoc](./doc/rationale1.adoc)
-- Rationale for implementing the steps for deriving a complete set of dimensionless products; [reStructuredText](./doc/rationale2.rst), [AsciiDoc](./doc/rationale2.adoc)
+- Tutorial: Generate dimensional formulae and perform consistency checking; [AsciiDoc](./doc/tutorial1.adoc), [reStructuredText for Bitbucket](./doc/tutorial1.rst)
+- Tutorial: Derive dimensionless products; [AsciiDoc](./doc/tutorial2.adoc), [reStructuredText for Bitbucket](./doc/tutorial2.rst)
+- Example: Journal Bearing; [AsciiDoc](./doc/tutorial3.adoc), [reStructuredText for Bitbucket](./doc/tutorial3.rst)
+- Rationale for the seven base dimensions; [AsciiDoc](./doc/rationale1.adoc), [reStructuredText for Bitbucket](./doc/rationale1.rst)
+- Rationale for implementing the steps for deriving a complete set of dimensionless products; [AsciiDoc](./doc/rationale2.adoc), [reStructuredText for Bitbucket](./doc/rationale2.rst)
 - [Source code documentation](https://cljdoc.org/d/com.neuralgraphs/diman) 
 
 ## License
