### 2021-10-04 #3735 @khinsen
Review update: I am looking at the code and playing with it. Interacting with the author via [issues](https://github.com/neuralgraphs/diman/issues?q=is%3Aissue) on the code repository.
