### 2021-10-04 6642b7f
corrected dimension symbol for diameter m in the table
