+++ doc/tutorial3.rst
@@ -48,7 +48,7 @@ Since,
 +-----------------+---------------+----------------+------------+
 | |M|             | moment        | |Nm|           | |ML2byT2|  |
 +-----------------+---------------+----------------+------------+
-| |D|	          | diameter      | |m|            | |M|        |
+| |D|	          | diameter      | |m|            | |L|        |
 +-----------------+---------------+----------------+------------+
 | |mu|            | viscosity     | |Pas|          | |MbyLT|    |
 +-----------------+---------------+----------------+------------+
