+++ doc/tutorial3.adoc
@@ -194,13 +194,13 @@ step such that it returns the solution matrix.
 == 4. Get Dimensionless Products
 
 ....
-=> (def all-dimless (get-dimensionless-products solution_matrix varpars2))
+=> (def all_dimless (get-dimensionless-products solution_matrix varpars2))
 
-=> (pprint all-dimless)
+=> (pprint all_dimless)
   [{:symbol "pi0", :expression "P^(1)*mu^(-1)*N^(-1)"}
    {:symbol "pi1", :expression "M^(1)*D^(-3)*mu^(-1)*N^(-1)"}]
 
-=> (get-pi-expression all-dimless "pi0")
+=> (get-pi-expression all_dimless "pi0")
   "P^(1)*mu^(-1)*N^(-1)"
 ....
 
