### 2021-10-15 #3735 @oyvinht
**Background**
Diman is a software package for dimension analysis of symbolic formulas and equations. The package basics include the necessary mechanisms for defining these constructs in a programming language (Clojure). All of the seven SI units for dimensions are covered. The gist of the package is tools for automated derivation of dimensionless products of formulae and checking for dimensional homogeneity of equations.

**Relevance**
The package would provide useful for anyone wanting to simplify or verify dimensional expressions (such as in engineering or science or building tools for these users) or possibly as a code basis for adding similar features to other automated symbolic processing systems (such as GNU Octave). From reading the statement of need, it also appears that the package is very likely to be cited as a tool in future publications from research in the field of dimensional analysis.

**Code guidelines**
Code builds and runs. There are a number of tests provided. The size of the project points to substantial work done by the author (apparently one person). The following notes should be taken:

1) Tests do not cover all of the code.

2) src/diman/analyze.clj:
   This file includes code that make a good example of a type of cleanup that the package could benefit from. The function "clean-dimnames" builds on three other functions "remove-empty-string", "replace-empty-string-by-nil" and "replace-plus-by-empty-string". For succinctness I believe it would be better to replace the first of these with the built-in "(remove empty? ..." (though the resulting list would be reversed). The second function seems to be a step that could be removed by just by removing 'nil instead of empty strings. Then again, the "replace-plus-by-empty-string" is (currently) just used this one place, so it appears unclear if there is any reason not to just reduce the whole procedure to "(remove empty? (clojure.string/split dimnames #"\+| "))".

3) src/diman/dimensions.clj:
   It is the reviewers _opinion_ that the strings used to identify units should be replaced by symbols in the diman.dimensions namespace. Development tools would benefit from this (e.g. code completion) and typos could easily be identified by the compiler, making the tool easier to use and more robust. It would also allow for expressing formulae as standard s-expressions, pertaining to the beneficial homoiconicity that Lisps provide. It would also allow for the use of the excellent symbol metadata features of Clojure.

4) The authors might consider utilizing (reader?) macros to provide an easier way of defining formulae and equations. As an example, I think I (and many other Clojure users) would enjoy being able to generate an equation with something like:

(with-variables (E dim/ENERGY
                 m dim/MASS
		 v dim/VELOCITY)
  (= E (* 1/2 m v v)))

Summary:
I recommend _accepting_ the article given that point 1 is amended, and that a general clean up is done on the code, so as to heighten the quality of the published code, and avoid unnecessary complexity "bugs" (as exemplified by 2 above) in an otherwise well crafted package.

Points 3 and 4 above are more about "taste", but it is my belief that most Clojure users would consider the code more idiomatic if they were adhered to.
