### 2021-10-15 c500a2e
get-solution-matrix replaces get-solved-matrix
