+++ doc/tutorial2.adoc
@@ -177,7 +177,7 @@ Size -> 3 x 4
 === 3.3. Get the solution matrix
 
 ....
-=> (view-matrix (get-solved-matrix (solve (get-augmented-matrix (generate-dimmat varpars2)))))
+=> (view-matrix (get-solution-matrix (solve (get-augmented-matrix (generate-dimmat varpars2)))))
 [1 0 -1/3 -1/4]
 [0 1 2N -7/4]
 Size -> 2 x 4
@@ -190,7 +190,7 @@ We can put all these individual steps involving matrix into one coding
 step such that it returns the solution matrix.
 
 ....
-=> (def solution_matrix (get-solved-matrix
+=> (def solution_matrix (get-solution-matrix
                             (solve
                                 (get-augmented-matrix
                                     (generate-dimmat varpars2)))))
