+++ doc/tutorial3.adoc
@@ -168,7 +168,7 @@ We add the above dimensional formulae into the
 === 3.3. Get the solution matrix
 
 ....
-=> (view-matrix (get-solved-matrix (solve (get-augmented-matrix (generate-dimmat varpars2)))))
+=> (view-matrix (get-solution-matrix (solve (get-augmented-matrix (generate-dimmat varpars2)))))
   [1 0 0N -1N -1N]
   [0 1 -3N -1N -1N]
   Size -> 2 x 5
@@ -181,7 +181,7 @@ We can put all these individual steps involving matrix into one coding
 step such that it returns the solution matrix.
 
 ....
-=> (def solution_matrix (get-solved-matrix
+=> (def solution_matrix (get-solution-matrix
                             (solve
                                 (get-augmented-matrix
                                     (generate-dimmat varpars2)))))
