### 2021-10-15 75b97a8
get-solution-matrix replaces get-solved-matrix
