### 2021-10-15 lungsi/diman#5 @lungsi
Clean the private functions called by `consistent?`
This issue was created following the review by @oyvinht

> @whedon @lungsi @mjsottile Here is my review:
>
> 2) src/diman/analyze.clj:
>    This file includes code that make a good example of a type of cleanup that the package could benefit from. The function "clean-dimnames" builds on three other functions "remove-empty-string", "replace-empty-string-by-nil" and "replace-plus-by-empty-string". For succinctness I believe it would be better to replace the first of these with the built-in "(remove empty? ..." (though the resulting list would be reversed). The second function seems to be a step that could be removed by just by removing 'nil instead of empty strings. Then again, the "replace-plus-by-empty-string" is (currently) just used this one place, so it appears unclear if there is any reason not to just reduce the whole procedure to "(remove empty? (clojure.string/split dimnames #"\+| "))".
>
> _Originally posted by @oyvinht in https://github.com/openjournals/joss-reviews/issues/3735#issuecomment-944203463_
