### 2021-10-15 lungsi/diman#7 @lungsi
Testing covers major parts of the source code
Added comprehensive tests for functions in `analyze.clj`. They accommodates _changes made for `consistent?` by @lungsi in https://github.com/neuralgraphs/diman/issues/5#issue-1027748371_ Some of the previous tests (other tests) were also cleaned.

> @whedon @lungsi @mjsottile Here is my review:
>
> 1) Tests do not cover all of the code.
>
> _Originally posted by @oyvinht in https://github.com/openjournals/joss-reviews/issues/3735#issuecomment-944203463_

With this the testing now covers major parts of the source code, only general utility functions and functions for accessing matrices (in `linalg` subpacakge) do not have tests.

Functions that currently don't have a test is because they are very simple functions. Also, although the coding for diman loosely began with a test-driven approach the current plan is to implement [clojures.spec](https://clojure.org/guides/spec)  as part of the diman code base. Thus, future codes will have both tests and specs.
