### 2021-10-15 #3735 @oyvinht
> @oyvinht Sorry about that - occasionally reviewers aren’t automatically assigned correctly in the review issue even if their checklist is present. I believe you’re now assigned and can edit it.

Thank you, it works well now!
