+++ CONTRIBUTING.adoc
@@ -1,0 +1,109 @@
+= Contributing to diman
+Lungsi Ngwua <aneuralgraphs@gmail.com>
+3.0, July 29, 2022: AsciiDoc article template
+:toc:
+:icons: font
+:url-quickref: https://docs.asciidoctor.org/asciidoc/latest/syntax-quick-reference/
+
+This article provides guidelines on how to contribute to diman.
+
+== What is an Issue?
+
+Contribution to the project is based on GitHub issues. Extensive use of labels categorizes the issues. Some "standard labels" to find issues and begin working are:
+
+* Something isn't working https://github.com/neuralgraphs/diman/labels/bug[[#d73a4a]#bug#]
+* Improvements or additions to documentation https://github.com/neuralgraphs/diman/labels/documentation[[#0075ca]#documentation#]
+* This issue or pull request already exists https://github.com/neuralgraphs/diman/labels/duplicate[[#cfd3d7]#duplicate#]
+* New feature or request https://github.com/neuralgraphs/diman/labels/enhancement[[#a2eeef]#enhancement#]
+* Good for newcomers https://github.com/neuralgraphs/diman/labels/good%20first%20issue[[#7057ff]#good first issue#]
+* Extra attention is needed https://github.com/neuralgraphs/diman/labels/help%20wanted[[#008672]#help wanted#]
+* This doesn't seem right https://github.com/neuralgraphs/diman/labels/invalid[[#e4e669]#invalid#]
+* Further information is requested https://github.com/neuralgraphs/diman/labels/question[[#d876e3]#question#]
+* Improve design/structure without changing behavior https://github.com/neuralgraphs/diman/labels/refactor[[#ccff00]#refactor#]
+* Make the code cleaner and clearer https://github.com/neuralgraphs/diman/labels/simplify[[#96c8a2]#simplify#]
+* This needs a test https://github.com/neuralgraphs/diman/labels/test[[#ff8c00]#test#]
+* This will not be worked on https://github.com/neuralgraphs/diman/labels/wontfix[[#ffffff]#wontfix#]
+
+=== Proposing a New Issue
+
+If a contributor has discovered a "potential issue" that has not yet been highlighted, follow the steps:
+
+. Create a GitHub issue with all the relevant details; the problem/issue and rationale for why it is a problem/issue and the proposed solution/change.
+* *Do not* make a pull request without creating the issue.
+. Wait for the created issue to be designated by a label.
+. Begin working on the labeled issue -- with some exceptions like https://github.com/neuralgraphs/diman/labels/wontfix[[#ffffff]#wontfix#], see above list.
+* See <<the_contribution_process,the contribution process>>
+
+==== A Note on Proposing New Issues
+
+Before posting https://github.com/neuralgraphs/diman/labels/question[[#d876e3]#question#] issue please make sure the exact or related question is not already part of the https://github.com/neuralgraphs/diman/discussions[discussions].
+
+== The Contribution Process [[the_contribution_process]]
+
+Steps to contribute are:
+
+. Comment on the labeled issue.
+* This avoids conflicts with another contributor who might work on the issue.
+. Write the code.
+* *All codes must* be <<documentation_guidelines, documented.>>
+* *All codes must* implement https://clojure.org/guides/spec[clojure.spec]footnote:[As of the time of writing this manual the diman codes are not yet specified, but they will be specified in the future and regardless of this situation all new contributing codes are expected to have its specification.]
+* Wherever appropriate there should be accompanying tests.
+. Submit a <<pull_request_guidelines, pull request.>>
+. Wait for code review and address any associated issue.
+
+=== Documentation Guidelines [[documentation_guidelines]]
+
[…]
