### 2021-10-16 neuralgraphs/diman#8 @lungsi
Added Contribution Guidelines
Following recommendations by @khinsen I have added [guidelines](https://github.com/neuralgraphs/diman/blob/master/CONTRIBUTING.adoc) for contributing to diman.

> @lungsi @mjsottile Here comes my review!
>
> A related point is the absence of community guidelines. It would be trivial for the author to add some boilerplate along the lines of "contribute via GitHub features". But the real point is that this package doesn't have a community and doesn't need one. It's small enough to be understood by a single person, even someone else than the author. I wouldn't mind using it for a research project even if I knew for sure there were no support at all. But it is nevertheless a useful contribution to science, because it is much less effort to use this package than to develop my own. In fact, I believe that packages such as this should be published because they make underappreciated techniques more accessible.
>
> _Originally posted by @khinsen in https://github.com/openjournals/joss-reviews/issues/3735#issuecomment-938733157_
