### 2021-10-18 neuralgraphs/diman#9 @lungsi
Response concerning "State of the field" for JOSS
> @lungsi @mjsottile Here comes my review!
>
> The final unchecked point is "state of the field". I am now aware of any similar package, and assuming that @lungsi doesn't either, he/she could simply add a sentence stating this. But that misses the point that the main reason for publishing this package is the lack of any other tools that implement the same techniques.
>
> _Originally posted by @khinsen in https://github.com/openjournals/joss-reviews/issues/3735#issuecomment-938733157_

> @khinsen @lungsi : Concerning "State of the field". I cannot find any published libraries on dimensional analysis in Clojure, but the problem is discussed online*. A book that keeps coming up is "Functional programming and applications" (ISBN: 10: 0133315797). I have not read that book, but it apparently covers solving the problem using functional languages (obviously not in Clojure, as the book dates back to 1980). It is possibly relevant for the "state of the field", though.
>
> * http://www.leadlab.ca/2016/07/27/solving-the-trivial-problem-of-dimensional-analysis-in-clojure/
> * https://github.com/mfey/units2 (somewhat)
>
> _Originally posted by @oyvinht in https://github.com/openjournals/joss-reviews/issues/3558#issuecomment-945703201_

Comments on the "State of the field" is given in the later paragraphs under the section "Statement of need".

> Numerous softwares have been developed to deal with dimensions in some shape or form ([Preussner, 2018](https://gmpreussner.com/research/dimensional-analysis-in-programming-languages); [Sharma, 2021](https://lungsi-slides.github.io/dataclub/icn/2021July26/index.html#/5/2)). Most incorporate the ability to tag quantities with units, however, few are capable of doing consistency checks and fewer still deal with dimensionless products let alone, deriving dimensionless products.
>
> diman is designed with an emphasis on analysis; the application of the algebraic theory of dimensionally homogeneous functions (Langhaar, 1951). It can check for dimensional homogeneity of a given equation and can derive the complete set of dimensionless products of a given equation.

In addition to those mentioned above by @oyvinht I have found [55 dimensional related software products](https://lungsi-slides.github.io/dataclub/icn/2021July26/index.html#/5/2) (as cited above (Sharma, 2021)&mdash;obsolete and active ones. The Clojure based units2 (https://github.com/mfey/units2 pointed out by @oyvinht) advertises to be "unit-aware computing", i.e. numerical computation with unit annotations, most of the discovered softwares target this use. Those that have analytical features are typically limited to performing checks for dimensional homogeneity. The number of tools capable of consistency checks are relatively lesser than those that target unit-aware computing. Tools capable of performing consistency checks are advertised as dimensional analysis tools.

Although the task of consistency checking with a pen and paper can be regarded as trivial (as suggested by the [link](http://www.leadlab.ca/2016/07/27/solving-the-trivial-problem-of-dimensional-analysis-in-clojure/) provided by @oyvinht), its software implementation is more involved, as illustrated by Peter Henderson in eleven pages in the book "Functional programming and applications" (ISBN: 0133315797), pages 60-71. Also many of these "dimensional analysis" tools that are capable of consistency checks perform consistency checks of units (e.g. if the right hand side has the unit feet/seconds and the left hand side has the unit meter/seconds, then the consistency check returns false), which suggests dimensions and units are considered as identical, but this is wrong. A unit reflects a particular measurement while a dimension is mathematical idea that refers to "the relationship between necessary coordinate concepts and the concept (of the quantity) which stands under them," the relationship and consequently the dimension is mathematically represented by some tag, in diman these are [M], [L], [T], [A], [K], [cd] and [mol].

The elementary function of dimensional checking&mdash;consistency checks of dimensions and not units&mdash;is a necessary function for derving dimensionless products. **This capability and the ability to derive complete set of dimensionless products differentiates diman from all the above mentioned softwares**. Although some dimensional softwares deal with dimensionless products in terms of a dimensionless unit tag for unit-aware computating, tools that deal with deriving dimensionless products are scarce. MATLAB&reg; provides [dimensional analysis feature](https://www.mathworks.com/discovery/dimensional-analysis.html) within the Symbolic Math Toolbox, but this does not include function(s) to derived dimensionless products; this ability may be acquired via an add-on script [(Thomas Tresch, 2021)](https://www.mathworks.com/matlabcentral/fileexchange/40890-buckinghams-pi-theorem-in-matlab).
