### 2021-10-20 neuralgraphs/diman#9 @khinsen
Your summary looks correct to me. There are indeed lots of solutions for unit-aware computation with physical quantities, for just about every language I know about, but that's something quite different from what `diman` does.
