### 2021-12-22 #3735 @mjsottile
@lungsi I've finished my review and it looks like you have addressed the reviewers.  I have the following final task for you to complete for us to push the submission to the final phase for acceptance.

The paper has a few very minor typos to correct:

- line 6 : dimensional formula -> dimensional formulas
- line 6 : dimensional equation -> dimensional equations
- line 10 : performs -> perform
- line 34 : softwares -> software packages
- line 43: comma after "Units" and before "diman".
- line 51: missing period at the end of the sentence.
- line 110: it might help to either add a small amount of horizontal spacing between each of the $\pi_N$ equations to separate them better, or typeset them to be on their own lines.
- line 112: you don't need to point out that pi doesn't denote the constant 3.14...  the meaning of pi in context is clear from the definition of the products above.
- line 119: do not often -> does not often

Regarding related work:

- You should include Kennedy's work in unit systems in your discussion of related work.  His work gets referenced pretty frequently by people implementing unit systems, and is the basis for units of measure in the F# language from Microsoft.  See: "A. J. Kennedy. Programming Languages and Dimensions. PhD thesis, University of Cambridge, 1995."  You can find papers that he's written that are shorter than his PhD thesis.

Once you have addressed these points, the final steps are:

- Please create an archive on zenodo or figshare for the submission, and post the DOI for the archive in this thread.
- Ensure that the version for the JOSS submission is correct.
- Ensure that the author list in the paper and archive are correct.

Once you've provided that information I can issue the final commands to pass the submission to an EIC for final processing.
