+++ paper.bib
@@ -63,3 +63,14 @@
   journal = {GitHub repository},
   url = {https://github.com/neuralgraphs/diman}
 }
+
+@techreport{Kennedy:1996,
+  author       = {Andrew John Kennedy}, 
+  title        = {Programming languages and dimensions},
+  institution  = {University of Cambridge},
+  year         = 1996,
+  number       = 391,
+  address      = {15 JJ Thomson Avenue, Cambridge CB3 0FD, United Kingdom},
+  month        = 4,
+  url         = {https://www.cl.cam.ac.uk/techreports/UCAM-CL-TR-391.pdf}
+}
