### 2022-01-05 78285cd
added Frink's presentation in bib
