+++ paper.bib
@@ -74,3 +74,10 @@
   month        = 4,
   url         = {https://www.cl.cam.ac.uk/techreports/UCAM-CL-TR-391.pdf}
 }
+
+@inproceedings{Eliasen:2004,
+    author  = {Alan Eliasen},
+    title   = {Frink - A Language for Understanding the Physical World},
+    publisher = {Lightweight Languages 4 (LL4) conference at MIT in December 2004},
+    year    = 2004
+}
