+++ paper.bib
@@ -79,5 +79,6 @@
     author  = {Alan Eliasen},
     title   = {Frink - A Language for Understanding the Physical World},
     publisher = {Lightweight Languages 4 (LL4) conference at MIT in December 2004},
-    year    = 2004
+    year    = 2004,
+    url = {https://frinklang.org/LL4.html}
 }
