### 2022-01-05 cf8223f
added corrections suggested by Matthew Sottile of JOSS
