### 2022-01-11 d10e137
added Kennedy's 2010 lecture note
