+++ paper.bib
@@ -75,6 +75,22 @@
   url         = {https://www.cl.cam.ac.uk/techreports/UCAM-CL-TR-391.pdf}
 }
 
+@Inbook{Kennedy:2010,
+  author = {Kennedy, Andrew John},
+  editor = {Horv{\'a}th, Zolt{\'a}n and Plasmeijer, Rinus and Zs{\'o}k, Vikt{\'o}ria},
+  title = {Types for Units-of-Measure: Theory and Practice},
+  bookTitle = {Central European Functional Programming School: Third Summer School, CEFP 2009, Budapest, Hungary, May 21-23, 2009 and Kom{\'a}rno, Slovakia, May 25-30, 2009, Revised Selected Lectures},
+  year = {2010},
+  publisher = {Springer Berlin Heidelberg},
+  address = {Berlin, Heidelberg},
+  pages = {268--305},
+  abstract = {Units-of-measure are to science what types are to programming. In science and engineering, dimensional and unit consistency provides a first check on the correctness of an equation or formula, just as in programming the validation of a program by the type-checker eliminates one possible reason for failure.},
+  isbn = {978-3-642-17685-2},
+  doi = {10.1007/978-3-642-17685-2_8},
+  url = {https://doi.org/10.1007/978-3-642-17685-2_8}
+}
+
+
 @inproceedings{Eliasen:2004,
     author  = {Alan Eliasen},
     title   = {Frink - A Language for Understanding the Physical World},
