+++ paper.md
@@ -35,7 +35,11 @@ If possible, it is beneficial to use the transformed parent function, where all 
 
 Numerous software packages have been developed to deal with dimensions in some shape or form [@Preussner:2018; @Sharma:2021]. Most incorporate the ability to tag quantities with units, however, few are capable of doing consistency checks and fewer still deal with dimensionless products let alone, deriving dimensionless products.
 
-`diman` is designed with an emphasis on **analysis**; the  application of the algebraic theory of dimensionally homogeneous functions [@Langhaar:1951]. It can check for dimensional homogeneity of a given equation and can derive the complete set of dimensionless products of a given equation.
+Frink and F sharp are two active languages that incorporates unit of measurement. Frink is a calculating tool and programming language designed to make physical calculations as it tracks unit of measure throughout calculations [Eliasen:2004]. F# is a general purpose language which allows floats and integers annotation with statically-typed unit metadata. F#'s units of measure is based on a prototype Meta Language (ML) like language that has dimensional type; the language supports type polymorphism as well as dimension polymorphism [Kennedy:1996].
+
+In F# the notion of dimension include ideas of type theory; therefore, dimensions satisfy algebraic properties of Abelian groups (non-commutative groups). The concepts of dimensional type and dimensional invariance can be mathematically extended to apply Buckingham's Theorem (also called Pi theorem) [Kennedy:1996]&mdash;this theorem is the basis for deriving the set of dimensionless products [@Ngwua:2020].
+
+`diman` is designed with an emphasis on **analysis**; the  application of the algebraic theory of dimensionally homogeneous functions [@Langhaar:1951]. Unlike F#'s units of measure which performs dimensional checks during calculations, checks for dimensional homogeneity in diman is manual. Also, the possibility of applying Buckingham's theorem in F# is done by mathematical abstraction [@Kennedy:2010] while diman provides functions to derive the complete set of dimensionless products of a given equation. Therefore, diman is easily put into practice for providing complete information of experimental results of physical systems in a compact form and also transform hypothetical function that describes the physical system [@Sharma:2021].
 
 # Design and implementation
 
