### 2022-01-11 #3735 @lungsi
@mjsottile Thank you so much for the feedback. Below are all the changes made in the revised version.

* [x] line 6 : dimensional formula -> dimensional formulas
* [x] line 6 : dimensional equation -> dimensional equations
* [x] line 10 : performs -> perform
* [x] line 34 : softwares -> software packages
* [x] line 43: comma after "Units" and before "diman".
   - line 60 in the revised version
* [x] line 51: missing period at the end of the sentence.
   - line 68 in the revised version
* [x] line 110: it might help to either add a small amount of horizontal spacing between each of the $\pi_N$ equations to separate them better, or typeset them to be on their own lines.
   - line 127 in the revised version
* [x] line 112: you don't need to point out that pi doesn't denote the constant 3.14... the meaning of pi in context is clear from the definition of the products above.
   - line 129 in the revised version
* [x] line 119: do not often -> does not often
   - line 135 in the revised version
* [x] You should include Kennedy's work in unit systems in your discussion of related work...
   - lines 38 to 58

In addition I have also done the following.

* [x] Please create an archive on zenodo or figshare for the submission, and post the DOI for the archive in this thread.
   - [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5837630.svg)](https://doi.org/10.5281/zenodo.5837630)
* [x] Ensure that the version for the JOSS submission is correct.
   - **v1.1.0** should be the version for JOSS submission.
* [x] Ensure that the author list in the paper and archive are correct.
