### 2022-01-12 #3735 @mjsottile
@lungsi Excellent - I think the added text talking about related work helps a lot in positioning your work with respect to other projects.  There are just a couple of grammatical issues in the new text.  Otherwise, it looks like we're good to go.

- line 38 : incorporates unit -> incorporate units
- line 40 : unit of measure -> units of measure
- line 41 : which allows floats and integers annotation with statically-typed unit metadata -> which allows annotating floating point and integer values with statically-typed unit metadata
- line 45 : include ideas of type theory -> is based on ideas from type theory
