### 2022-01-12 #3735 @mjsottile
@lungsi Also, the bibliography entry for Kennedy (2010) needs to be fixed.  The capitalization of place names is incorrect (the city and country names should be capitalized, as well as the name "European").  I am assuming this is a side effect of BiBTeX or LaTeX formatting your original text when generating the PDF.  You should put curly braces around capital letters to force them to be upper case in the rendered output.  Let me know if you have trouble getting the capitalization fixed for that entry.
