+++ paper.bib
@@ -39,7 +39,7 @@
 @misc{Ngwua:2020,
 	title = {Theory of Dimensionless Products},
 	url = {https://www.neuralgraphs.com/lectures/diman/lectp1.html},
-	abstract = {Lectures on the Theory of Dimensionless Products},
+	abstract = {Lectures on the {Theory} of {Dimensionless} {Products}},
 	date = {2020-07-25},
 	journal = {Neuralgraphs},
 	author = {Ngwua, Lungsi},
@@ -50,7 +50,7 @@
 @online{Clojure:2020,
   	url = {https://clojure.org/reference/data_structures},
   	Author = {Hickey, Rich},
-  	title = {Clojure - Data Structures},
+  	title = {Clojure - {Data} {S}tructures},
   	organization = {Clojure},
   	Year = 2021,
 }
@@ -78,8 +78,8 @@
 @Inbook{Kennedy:2010,
   author = {Kennedy, Andrew John},
   editor = {Horv{\'a}th, Zolt{\'a}n and Plasmeijer, Rinus and Zs{\'o}k, Vikt{\'o}ria},
-  title = {Types for Units-of-Measure: Theory and Practice},
-  bookTitle = {Central European Functional Programming School: Third Summer School, CEFP 2009, Budapest, Hungary, May 21-23, 2009 and Kom{\'a}rno, Slovakia, May 25-30, 2009, Revised Selected Lectures},
+  title = {Types for {Units-of-Measure}: {Theory} and {Practice}},
+  bookTitle = {Central {European} {Functional} {Programming} {School}: {Third} {Summer} {School}, CEFP 2009, {Budapest}, {Hungary}, May 21-23, 2009 and {K}om{\'a}rno, {S}lovakia, May 25-30, 2009, {Revised} {Selected} {L}ectures},
   year = {2010},
   publisher = {Springer Berlin Heidelberg},
   address = {Berlin, Heidelberg},
@@ -93,7 +93,7 @@
 
 @inproceedings{Eliasen:2004,
     author  = {Alan Eliasen},
-    title   = {Frink - A Language for Understanding the Physical World},
+    title   = {Frink - {A} {Language} for {Understanding} the {Physical} {World}},
     publisher = {Lightweight Languages 4 (LL4) conference at MIT in December 2004},
     year    = 2004,
     url = {https://frinklang.org/LL4.html}
