+++ paper.bib
@@ -79,7 +79,7 @@
   author = {Kennedy, Andrew John},
   editor = {Horv{\'a}th, Zolt{\'a}n and Plasmeijer, Rinus and Zs{\'o}k, Vikt{\'o}ria},
   title = {Types for {Units-of-Measure}: {Theory} and {Practice}},
-  bookTitle = {Central {European} {Functional} {Programming} {School}: {Third} {Summer} {School}, CEFP 2009, {Budapest}, {Hungary}, May 21-23, 2009 and {K}om{\'a}rno, {S}lovakia, {M}ay 25-30, 2009, {Revised} {Selected} {L}ectures},
+  bookTitle = {Central {European} {Functional} {Programming} {School}: {Third} {Summer} {School}, CEFP 2009, {Budapest}, {Hungary}, {May} 21-23, 2009 and {K}om{\'a}rno, {S}lovakia, {May} 25-30, 2009, {Revised} {Selected} {L}ectures},
   year = {2010},
   publisher = {Springer Berlin Heidelberg},
   address = {Berlin, Heidelberg},
