### 2022-01-13 #3735 @lungsi
@mjsottile Thank you very much. I have made the changes as suggested.

* [x] line 38 : incorporates unit -> incorporate units
* [x] line 40 : unit of measure -> units of measure
* [x] line 41 : which allows floats and integers annotation with statically-typed unit metadata -> which allows annotating floating point and integer values with statically-typed unit metadata
* [x] line 45 : include ideas of type theory -> is based on ideas from type theory
* [x] the bibliography entry for Kennedy (2010) needs to be fixed ...
