+++ paper.bib
@@ -37,7 +37,7 @@
 }
 
 @misc{Ngwua:2020,
-	title = {Theory of Dimensionless Products},
+	title = {Theory of {Dimensionless} {Products}},
 	url = {https://www.neuralgraphs.com/lectures/diman/lectp1.html},
 	abstract = {Lectures on the {Theory} of {Dimensionless} {Products}},
 	date = {2020-07-25},
