### 2022-01-13 4fc1d59
minor paper updates
