### 2022-01-13 #3735 @danielskatz
@lungsi - I've suggested some small changes in https://github.com/neuralgraphs/diman/pull/10 - please merge this or let me know what you disagree with.
