### 2022-01-13 #3735 @lungsi
> @lungsi - I've suggested some small changes in [neuralgraphs/diman#10](https://github.com/neuralgraphs/diman/pull/10) - please merge this or let me know what you disagree with.

I have merged it. Thanks @danielskatz
