### 2022-01-14 #3735 @lungsi
Thank you very much (reviewers: @oyvinht and @khinsen, and editors: @mjsottile and @danielskatz).

Comments by @oyvinht and @khinsen were very helpful, I learned some new stuffs. I really appreciated the generous comments from @khinsen, it was very encouraging. I would also like to point out that @mjsottile 's feedback greatly improved the manuscript.

Overall, my experience with JOSS has been highly rewarding (regardless of the paper being accepted, although this helps :) ).
